spinner lock on view wip
