+++ test-engine/src/ui/views/indicators/mod.rs
@@ -1,3 +1,5 @@
 mod spinner;
+mod spinner_locks;
 
 pub use spinner::*;
+pub use spinner_locks::*;
